+++ src/FileSystemProviders.ts
@@ -2,49 +2,86 @@ import axios from 'axios';
 import * as vscode from 'vscode';
 
 export interface FileSystemProvider {
-	getSize(uri: vscode.Uri): Promise<number | undefined>
-	getChildren(uri: vscode.Uri): Promise<FileItem[] | undefined>;
+	getSize(uri: vscode.Uri): number | undefined
+	getChildren(uri: vscode.Uri): FileItem[] | undefined;
+
+	fetch(uri: vscode.Uri): Promise<void>;
+	fetchSize(uri: vscode.Uri): Promise<void>;
+
 	update(): void;
 	init(rootUri: vscode.Uri): Promise<void>; 
 }
 
+interface FileItem {
+	uri: vscode.Uri;
+	isFolder: boolean;
+}
+
+interface FileItemVFS extends FileItem {
+	size?: number;
+	children: FileItem[];
+}
+
 export class VSCodeFileSystemProvider implements FileSystemProvider {
-	cache = new Map<string, number>();
+	cache = new Map<string, FileItemVFS>();
 
-	async init() {
-
+	async init(uri: vscode.Uri) {
 	}
 
-	async getChildren(uri: vscode.Uri) {
+	async fetch(uri: vscode.Uri) {
+		let children = await this.fetchChildren(uri);
+		let root = {
+			uri, isFolder: true, size: undefined, children
+		};
+		this.cache.set(uri.toString(), root);
+	}
+
+	async fetchSize(uri: vscode.Uri) {
+		let item = this.cache.get(uri.toString());
+		// if (!item) {
+		// 	await this.fetch(uri);
+		// 	item = this.cache.get(uri.toString())!;
+		// }
+		if (item?.isFolder) {return;}
+		item = item ?? {uri, isFolder: false, size: undefined, children: []};
+		let fsStat = await vscode.workspace.fs.stat(uri);
+
+		item.size = fsStat.size;
+		this.cache.set(uri.toString(), item);
+	}
+
+	async fetchChildren(uri: vscode.Uri): Promise<FileItem[]> {
 		return vscode.workspace.fs.readDirectory(uri).then((entries) => {
 			return entries.map(([name, type]) => {
 				const fUri = vscode.Uri.joinPath(uri, name);
 				return {
 					uri: fUri,
 					isFolder: type === vscode.FileType.Directory,
-					size: type === vscode.FileType.File ? -1 : 0
 				};
 			});
 		});
 	}
 
-	async getSize(uri: vscode.Uri): Promise<number | undefined> {
-		const key = uri.toString();
-		if (!this.cache.has(key)) {
-			let fsStat = await vscode.workspace.fs.stat(uri);
-			this.cache.set(key, fsStat.size);
+	getChildren(uri: vscode.Uri) {
+		let item = this.cache.get(uri.toString());
+		if (!item) {
+			return undefined;
 		}
-		return this.cache.get(key)!;
+		return item.children;
+	}
+
+	getSize(uri: vscode.Uri) {
+		const key = uri.toString();
+		let item = this.cache.get(key);
+		if (!item) {
+			return undefined;
+		}
+		return item.size;
 	}
 
 	update() {
 		this.cache = new Map();
 	}
-}
-
-interface FileItem {
-	uri: vscode.Uri;
-	isFolder: boolean;
 }
 
 interface GitObject {
@@ -68,7 +105,7 @@ export class GitHubFileSystemProvider implements FileSystemProvider {
 	fsTree: any = {};
 	rootUris?: vscode.Uri[];
 
-	async getChildren(uri: vscode.Uri) {
+	getChildren(uri: vscode.Uri) {
 		let gitObject = this.cache.get(uri.toString());
 		if (!gitObject) {
 			return undefined;
@@ -84,23 +121,38 @@ export class GitHubFileSystemProvider implements FileSystemProvider {
 		});
 	}
 
+	fetch(uri: vscode.Uri) {
+		return this.init(uri);
+	}
+
+	fetchSize(uri: vscode.Uri) {
+		return this.init(uri);
+	}
+
 	async init(uri: vscode.Uri) {
+		if (uri.scheme !== 'vscode-vfs') return;
+        if (!uri.authority.startsWith('github')) return;
+
 		let str = uri.toString();
 		if (!this.rootUris?.some(el => el.toString() === str)) {
 			this.rootUris?.push(uri);
 		}
 
 		await this._initFor(uri);
-
-		// this.cache = new Map([...this.cache.entries()].sort((a, b) => b[0].length - a[0].length));
 	}
 
 	async fetchRepoOf(uri: vscode.Uri) {
 		let meta: any = uri.authority.split(/(\+|%2B)/)[2];
-		meta = JSON.parse(hex2utf8(meta));
+		if (meta) {
+			meta = JSON.parse(hex2utf8(meta));
+		}
         let owner = uri.path.split('/')[1];
         let name = uri.path.split('/')[2];
-        let laxios = axios.create({ baseURL: "https://api.github.com" });
+		let headers = {};
+		if (this.token) {
+			headers = { Authorization: `token ${this.token}` } ;
+		}
+        let laxios = axios.create({ baseURL: "https://api.github.com", headers });
 		let oid;
 
 		uri = uri.with({path: `/${owner}/${name}`});
@@ -112,17 +164,14 @@ export class GitHubFileSystemProvider implements FileSystemProvider {
         } else if (meta.ref?.type === 2) {
 			oid = meta.ref.id;
 		} else if (meta.ref?.type === 3) {
-			oid = (await laxios.get(`/repos/${owner}/${name}/git/pulls/${meta.ref.id}`)).data.head.sha;
-		} else if (meta.ref?.type === 4) {
+			oid = (await laxios.get(`/repos/${owner}/${name}/pulls/${meta.ref.id}`)).data.head.sha;
+		} else {
             oid = (await laxios.get(`/repos/${owner}/${name}/git/ref/heads/${meta.ref.id}`, {
 				validateStatus: () => true
 			}))?.data?.object?.sha;
 			if (!oid) {
 				oid = (await laxios.get(`/repos/${owner}/${name}/git/ref/tags/${meta.ref.id}`))?.data?.object?.sha;
 			}
-		} else {
-			console.error("Size Extension: Couldn't determine ref for " + owner + '/' + name + ". Fallback to vscode API.");
-			return;
 		}
         let response = await laxios.get(`/repos/${owner}/${name}/git/trees/${oid}?recursive=1`);
         let trees = response.data.tree as GitObject[];
@@ -134,7 +183,7 @@ export class GitHubFileSystemProvider implements FileSystemProvider {
 			type: 'tree',
 			size: 0,
 			sha: oid,
-			url: uri.toString(),
+			url: uri.toString(),                                     
 			children: [],
 			folderSize: 0,
 			count: 0
@@ -159,6 +208,8 @@ export class GitHubFileSystemProvider implements FileSystemProvider {
         }
 	}
 
+	token: string | undefined;
+
 	async _initFor(uri: vscode.Uri): Promise<void> {
 		if (uri.scheme !== 'vscode-vfs') return;
         if (!uri.authority.startsWith('github')) return;
@@ -167,8 +218,15 @@ export class GitHubFileSystemProvider implements FileSystemProvider {
 			await this.fetchRepoOf(uri);
 			console.log("Size Extension: Folder " + uri.toString() + " will be handled with GitHub API.");
 		} catch(e: any) {
-			console.error("Size Extension: Couldn't fetch github repository " + uri.toString() + ". Fallback to vscode API.");
-			console.error(e);
+			if (e.response.status === 403) {
+				this.token = await getAuthToken().catch(() => undefined);
+				if (this.token) {
+					await this._initFor(uri);
+					return;
+				}
+			}
+			console.error("Size Extension: Couldn't fetch github repository " + uri.toString() + ", fallback to vscode API.", e);
+			return;
 		}
 
 		let entries = [...this.cache.entries()];
@@ -190,11 +248,11 @@ export class GitHubFileSystemProvider implements FileSystemProvider {
         rootUris.forEach((uri) => this._initFor(uri));
     }
 
-	async getSize(uri: vscode.Uri): Promise<number| undefined> {
+	getSize(uri: vscode.Uri) {
 		if (!this.cache.has(uri.toString())) {
 			return undefined;
 		}
-		return Promise.resolve(this.cache.get(uri.toString())!.size);
+		return this.cache.get(uri.toString())!.size;
 	}
 
 	update(): void {
@@ -204,4 +262,22 @@ export class GitHubFileSystemProvider implements FileSystemProvider {
 
 function hex2utf8(hexx: string) {
     return decodeURIComponent(hexx.replace(/\s+/g, '').replace(/[0-9a-f]{2}/g, '%$&'));
+}
+
+function getAuthToken(): Promise<string> {
+	return new Promise((resolve, reject) => {
+		vscode.authentication.getSession('github', ['repo'], { silent: true }).then( (e) => {
+			if (e) {
+				resolve(e.accessToken);
+				return;
+			}
+			vscode.window.showErrorMessage("Size: Anonymous Github API limit exceeded.", { modal: true, detail: "Use credentials to continue" }, { title: "Login via GitHub" }).then((e) => {
+				vscode.authentication.getSession('github', ['repo'], { createIfNone: true }).then( (e) => { 
+					resolve(e.accessToken);
+				}, (e) => {
+					reject();
+				});
+			 });
+		});
+	});
 }