+++ CHANGELOG.md
@@ -4,13 +4,32 @@ All notable changes to the "size" extension will be documented in this file.
 
 ## [Unreleased]
 
-## [0.0.11]
+## [0.2.1]
+
+- Minor tweaks and fixes
+
+## [0.2.0]
+
+- Added grouping by file extension
+- Reworked `View as` functionality
+- Reworked ignore feature, four separate pattern types
+
+## [0.1.23]
+
+- Added pattern ignore feature
+
+## [0.1.21]
+
+- Added `size.compactFolders` option (default: `true`)
+- Fixed .gitignore for folders
+
+## [0.1.11]
 
 ### Added
 
 - Experimental .gitignore support
 
-## [0.0.8]
+## [0.1.8]
 
 ### Added